+++ CHANGELOG.md
@@ -6,6 +6,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org).
 
+[v0.3.0]
+--------
+
+### Added
+- Raw, unprocessed parsed TaggedValues now trigger a ValueEvent named
+  `Symfony\Component\Yaml\Tag\TaggedValue-{tagName}`. After this event is
+  handled, the resulting value is processed and then fired in a separate
+  `{tagName}` ValueEvent.
+
+### Fixed
+- Inner values of parsed TaggedValue objects should be processed
+
 [v0.2.0]
 --------